# core: switch Bn256Add precompile to evmone